+++ src/views/containers/MenuBtn/index.ts
@@ -1,0 +1,21 @@
+import MenuBtn from '@atom/toggle/Menu';
+import { connect } from 'react-redux';
+
+const mapReduxStateToReactProps = ({ menuState }: any) => {
+  return { ...menuState };
+};
+
+const mapReduxDispatchToReactProps = (dispatch: any) => {
+  return {
+    onClick: () => {
+      dispatch({ type: 'TOGGLE_MENU' });
+    },
+  };
+};
+
+const connector = connect(
+  mapReduxStateToReactProps,
+  mapReduxDispatchToReactProps,
+);
+
+export default connector(MenuBtn);