+++ CHANGELOG.md
@@ -1,14 +1,25 @@
 # Change Log
 
 이 프로젝트에서 주목할만한 모든 변경 사항이이 파일에 문서화됩니다.
-[Keep a Changelog](https://keepachangelog.com/ko/1.0.0/)의 형식을 기본으로 구성됩니다.
-[Semantic Versioning](https://semver.org/lang/ko/) 프로젝트를 고수합니다.
+[Keep a Changelog](https://keepachangelog.com/ko/1.0.0/)의 형식을 기본으로 구성됩니다. 현재 버전은 [#181](https://github.com/Hansanghyeon/4log/issues/181)위와 같은 내용으로 구성됩니다.
+
+## [1.8.1] - 2020-06-06
+
+### Changed
+
+- [#181](https://github.com/Hansanghyeon/4log/issues/181) ~~[Semantic Versioning](https://semver.org/lang/ko/)~~ 프로젝트룰을 제거 독자적 버저닝으로 변경
+
+## [1.8.0] - 2020-06-02
+
+### Add
+
+- [#88](https://github.com/Hansanghyeon/4log/issues/88) 모바일 메뉴관련 버튼 추가 이전에는 모바일 md =< 786 사이즈 이하로 떨어질때 카테고리 메뉴를 완전 제거했는데 버튼을 통한 메뉴 보기 기능을 가하였다. `react-redux`를 활용 나중에 offset 밖영역 클릭도 리듀서를 통한 간단하게 기능추가가 가능함
 
 ## [1.7.0] - 2020-06-01
 
 ### Add
 
-- [[#172](https://github.com/Hansanghyeon/4log/issues/172) 4log 저장소에 적용한 코드에 대한 라이센스 적용
+- [#172](https://github.com/Hansanghyeon/4log/issues/172) 4log 저장소에 적용한 코드에 대한 라이센스 적용
 
 ### Changed